aws-generative-ai-for-documents-workshop-assets/stream6_working_with_databases/requirements.txt
streamlit 
mysql-connector-python 
pandas